agent-rag-workflow/pyproject.toml
[project]
name = "agent-rag-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=25.12.0",
    "chromadb>=1.4.0",
    "faiss-cpu>=1.13.2",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.15",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
